R3K1-07-Responsive-Web-Application/pyproject.toml
[project]
name = "repl-nix-physicsquizmaster"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.0",
    "sympy>=1.13.3",
]
